Attempts to fix small errors
After running a previous version of the script I have succeeded in spliting the ethanol use across the first 3 cols of the ethanol plate.  However Tris and Ethanol still aren't distributing to all 12 cols of the mag_plate, only to one (and different ones for the ethanol and tris steps!)

I have tried to fix a bug where only the first 2 cols of eluted sample were transfered to the output plate.